pyproject.toml
[tool.poetry]
name = "y-lab-fastapi"
version = "1"
description = "This is fastapi aplication"
authors = ["gglo0ol <124757163+gglo0ol@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10.13"
fastapi = "^0.109.0"
sqlalchemy = {extras = ["acincio"], version = "^2.0.25"}
uvicorn = "0.27.0"
psycopg2-binary = "2.9.9"
asyncpg = "0.29.0"
pydantic-settings = "2.1.0"
redis = "^5.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
